Client and server can now communicate
-Fire up nameserver : python -m Pyro4.naming
-Run the Server.
-Run The client_tests.

Server now registers itself with the nameserver.
You will see results show up in the server's window, from the client.
Notifier now hunts down the server within the name_server, and
reconnects to it if it drops. It attempts a test ping to ensure
connectivity.
TODO: Fix up client so it holds onto :
notifier/monitorManager/DataManager