Fixed the way that the datamanager returns it's keys.
Also, stopped the notifier from automatically firing off a connection on
instantiation, it makes testing much easier and provides no real
benefits.
On the server-side, modified the client-grabber to account for a fact
that when a client drops and reconnects, it will be served up a new URI.
This keeps our dictionary intact and accurate.