Wrote message handler closed #14
Wrote mock server objects for testing. Whenever testing=True on the
classes that require server instantiation, they receive mock server
objecst instead. The mock server was written into Utils.py.

Added a method in monitorManager that sends a results dictionary off to
the message handler for sending to the server.

Message handler does the following: Receives some message, appends the
client id, sends it to the message queue of the server.